Nodes that shall be loaded are managed in a global list. Fixed bug in octree tex generation.

+++ src/Serialization/OctantComponent.cs
@@ -53,7 +53,7 @@ namespace Fusee.Serialization
         /// </summary>
         public double ProjectedScreenSize { get; private set; }
 
-        public byte PosInHierarchyTex;
+        public int PosInHierarchyTex;
 
         public byte VisibleChildIndices;
